add label just below qrcode

+++ src/main/resources/assets/xuanran001.com/get-offline-fangan.sh
@@ -5,6 +5,7 @@ function get_one
   echo "get_one : $1"
   url="http://www.xuanran001.com/rwrule/zxfazzjfb3/${1}.html"
 
+  # zxfazzjfb3_db205959-2872-482a-af66-f6672ceb543d.html
   fangan_html="zxfazzjfb3_${1}.html"
 
   echo "<li><a href=\"${fangan_html}\">$fangan_html</a></li>" >> list.html
@@ -12,29 +13,54 @@ function get_one
   echo "download static file from $url"
   wget "$url" -O "$fangan_html" &> /dev/null
 
+  # Absolute path to relative path
+  sed -i "s/src=\"http:\/\/www\.xuanran001\.com\//src=\".\//g" $fangan_html
+  sed -i "s/href=\"http:\/\/www\.xuanran001\.com\//href=\".\//g" $fangan_html
+  sed -i "s/\"\/libs/\"\.\/libs/g" $fangan_html
+  sed -i "s/\"\/sites/\"\.\/sites/g" $fangan_html
+  sed -i "s/'\/sites/'\.\/sites/g" $fangan_html
+  sed -i "s/\"\/public/\"\.\/public/g" $fangan_html
+  sed -i "s/url(\/sites/url(.\/sites/g" $fangan_html
+
   # hide footer and header
   sed -i '1i<style>#header { padding : 0px !important; }#header .container, #footer { display : none !important; }body { padding: 0px !important; }#toolbar, #drupal_tabs { display: none !important; } body, div, a, small { color: black !important; }</style>' "$fangan_html"
+
+  # some changed requirement
+  cat << 'EOF' >> $fangan_html
+<script>
+$(function() {
+  $("h1").css("margin-bottom", "90px");
+  $("#mobile-qr").append('<br><span style="font-size: 14px;">扫一扫，带走全景效果图。</span>');
+});
+</script>
+EOF
 
   # gaibaojia
   echo "gaibaojia-start"
   sed -i 's/原地面清理，贴墙砖采用国际325#水泥以1:3配放砂浆，单价含水泥砂浆、人工，主材另算（规格为300mm\*300mm至800mm\*800mm）水泥沙厚度5cm以内，超过5cm另加10元\/平方米，斜铺另加8元\/平方 ，拼花另加30元\/平方米。/1、清理基层，扫除浮灰，洒水润湿；2、预排、放样，水泥砂浆铺贴；3、不含踢脚板安装、不含特殊基层处理、瓷砖规格：大于等于600mm小于等于800mm/g' "$fangan_html"
   echo "gaibaojia-end"
 
-  # get swf
-  # public/repository/108e/d900/8282/47d5/9597/a7c6/7b90/2938/swf
-  #swf_path=`curl "http://www.xuanran001.com/d/xanonymous?path=projects/schemeRoomList/id_${1}/test_1" 2> /dev/null | grep -Po '(?<="bigImg":")[^"]*' | grep "swf" | sed -e 's/\\\//g' | sed -e 's/\///' | sed -e 's/\/0000.swf//'`
-  swf_path=`grep -Po '(?<="imagePath":")[^"]*' "$fangan_html" | grep "swf" | sed -e 's/\\\//g' | sed -e 's/\///' | sed -e 's/\/0000.swf//'`
-  echo "swf path : $swf_path"
-  mkdir -p $swf_path
-  wget "http://www.xuanran001.com/$swf_path/0000.swf" -O "$swf_path/0000.swf" &> /dev/null
+  if [ $download_data = "1" ]; then
 
-  mkdir -p tmp
-  cd tmp
-  wget -r -p -np -k "$url" &> /dev/null
-  cp -rf www.xuanran001.com/public ../
-  cd ..
-  rm -rf tmp
+    # get swf
+    # public/repository/108e/d900/8282/47d5/9597/a7c6/7b90/2938/swf
+    swf_path=`grep -Po '(?<="imagePath":")[^"]*' "$fangan_html" | grep "swf" | sed -e 's/\\\//g' | sed -e 's/\///' | sed -e 's/\/0000.swf//'`
+    echo "swf path : $swf_path"
+    mkdir -p $swf_path
+    wget "http://www.xuanran001.com/$swf_path/0000.swf" -O "$swf_path/0000.swf" &> /dev/null
+
+    # get static file in /public
+    mkdir -p tmp
+    cd tmp
+    wget -r -p -np -k "$url" &> /dev/null
+    cp -rf www.xuanran001.com/public ../
+    cd ..
+    rm -rf tmp
+
+  fi
 }
+
+download_data="0"
 
 uuids=(
 db205959-2872-482a-af66-f6672ceb543d
@@ -56,9 +82,3 @@ do
   get_one "$uuid"
 done
 
-sed -i "s/src=\"http:\/\/www\.xuanran001\.com\//src=\".\//g" zxfazzjfb3_*
-sed -i "s/\"\/libs/\"\.\/libs/g" zxfazzjfb3_*
-sed -i "s/\"\/sites/\"\.\/sites/g" zxfazzjfb3_*
-sed -i "s/'\/sites/'\.\/sites/g" zxfazzjfb3_*
-sed -i "s/\"\/public/\"\.\/public/g" zxfazzjfb3_*
-sed -i "s/url(\/sites/url(.\/sites/g" zxfazzjfb3_*
